Cleans up the hack and is easier to use.
The previous hacked version was getting too painful to work with, as I
had to go into irb and require the file, then mess too much with the
Extractor. So I cleaned it up and made it so that I can just call the
prayer file with ruby. The extractor file is renamed from endings.rb
to extractor.rb.